+++ daemon/container/state.go
@@ -46,8 +46,8 @@ type State struct {
 	Health            *Health
 	Removed           bool `json:"-"`
 
-	stopWaiters       []chan<- container.StateStatus
-	removeOnlyWaiters []chan<- container.StateStatus
+	stopWaiters       []chan<- StateStatus
+	removeOnlyWaiters []chan<- StateStatus
 
 	// The libcontainerd reference fields are unexported to force consumers
 	// to access them through the getter methods with multi-valued returns
@@ -56,6 +56,26 @@ type State struct {
 
 	ctr  libcontainerdtypes.Container
 	task libcontainerdtypes.Task
+}
+
+// StateStatus is used to return container wait results.
+// Implements exec.ExitCode interface.
+// This type is needed as State include a sync.Mutex field which make
+// copying it unsafe.
+type StateStatus struct {
+	exitCode int
+	err      error
+}
+
+// ExitCode returns current exitcode for the state.
+func (s StateStatus) ExitCode() int {
+	return s.exitCode
+}
+
+// Err returns current error for the state. Returns nil if the container had
+// exited on its own.
+func (s StateStatus) Err() error {
+	return s.err
 }
 
 // NewState creates a default state object.
@@ -138,20 +158,23 @@ func (s *State) StateString() container.ContainerState {
 // be nil and its ExitCode() method will return the container's exit code,
 // otherwise, the results Err() method will return an error indicating why the
 // wait operation failed.
-func (s *State) Wait(ctx context.Context, condition container.WaitCondition) <-chan container.StateStatus {
+func (s *State) Wait(ctx context.Context, condition container.WaitCondition) <-chan StateStatus {
 	s.Lock()
 	defer s.Unlock()
 
 	// Buffer so we can put status and finish even nobody receives it.
-	resultC := make(chan container.StateStatus, 1)
+	resultC := make(chan StateStatus, 1)
 
 	if s.conditionAlreadyMet(condition) {
-		resultC <- container.NewStateStatus(s.ExitCode(), s.Err())
+		resultC <- StateStatus{
+			exitCode: s.ExitCodeValue,
+			err:      s.Err(),
+		}
 
 		return resultC
 	}
 
-	waitC := make(chan container.StateStatus, 1)
+	waitC := make(chan StateStatus, 1)
 
 	// Removal wakes up both removeOnlyWaiters and stopWaiters
 	// Container could be removed while still in "created" state
@@ -166,8 +189,10 @@ func (s *State) Wait(ctx context.Context, condition container.WaitCondition) <-c
 		select {
 		case <-ctx.Done():
 			// Context timeout or cancellation.
-			resultC <- container.NewStateStatus(-1, ctx.Err())
-
+			resultC <- StateStatus{
+				exitCode: -1,
+				err:      ctx.Err(),
+			}
 			return
 		case status := <-waitC:
 			resultC <- status
@@ -397,8 +422,11 @@ func (s *State) Err() error {
 	return nil
 }
 
-func (s *State) notifyAndClear(waiters *[]chan<- container.StateStatus) {
-	result := container.NewStateStatus(s.ExitCodeValue, s.Err())
+func (s *State) notifyAndClear(waiters *[]chan<- StateStatus) {
+	result := StateStatus{
+		exitCode: s.ExitCodeValue,
+		err:      s.Err(),
+	}
 
 	for _, c := range *waiters {
 		c <- result
