+++ daemon/wait.go
@@ -3,6 +3,7 @@ package daemon
 import (
 	"context"
 
+	"github.com/docker/docker/daemon/container"
 	containertypes "github.com/moby/moby/api/types/container"
 )
 
@@ -13,7 +14,7 @@ import (
 // condition is met or if an error occurs waiting for the container (such as a
 // context timeout or cancellation). On a successful wait, the exit code of the
 // container is returned in the status with a non-nil Err() value.
-func (daemon *Daemon) ContainerWait(ctx context.Context, name string, condition containertypes.WaitCondition) (<-chan containertypes.StateStatus, error) {
+func (daemon *Daemon) ContainerWait(ctx context.Context, name string, condition containertypes.WaitCondition) (<-chan container.StateStatus, error) {
 	cntr, err := daemon.GetContainer(name)
 	if err != nil {
 		return nil, err
